pyproject.toml
[tool.poetry]
name = "otterobt"
version = "0.1.0"
description = "Telegram Boardgame Assistant Bot"
authors = ["Reagan <rrr.chanhiulok@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.11.1"
pydantic = "^2.10.6"
openai = "^1.65.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.9"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
